requirements.txt
Django
psycopg2
djangorestframework
social-auth-app-django
google-api-python-client
django-widget-tweaks
